set core as the default page for the showcase
